Updated files before refactor
