+++ main.py
@@ -16,8 +16,7 @@ if torch.cuda.is_available():
     except ImportError:
         print("torch-tensorrt is not installed. Running on CPU mode only.")
 
-from benchmark.benchmark_utils import plot_benchmark_results
-from common.utils import parse_arguments
+from common.utils import parse_arguments, plot_benchmark_results
 from src.image_processor import ImageProcessor
 from src.model import ModelLoader
 import warnings
